feat: disable oauth2 client deletion
This new feature allows to disable hydra-maester OAuth2 client deletion.
It can be a pitfall that, when a Kube CRD gets deleted, the OAuth2 client gets deleted.

OAuth2 clients can be critical for production environments to work properly,
and allowing to delete one when a Kubernetes CRD gets deleted has shown to be brittle.

This feature should make hydra-maester safer.